+++ examples/delta_fofb.py
@@ -8,17 +8,21 @@ from os import makedirs
 
 from cobea.model import Response
 from cobea import cobea, read_elemnames
-from delta import drift_info
+from delta import drift_info, makedir_if_nonexistent
 import cobea.plotting as plt
 
 input_prefix = 'delta_fofb_input/'
 
 
-def import_response(filename='20_10_2017/fofb_rpm.txt'):
-    corr_labels = loadtxt(input_prefix + 'mag_list.csv', delimiter=',', usecols=(3, 4), dtype=str)
-
+def generate_corr_labels(mag_list='delta_fofb_input/mag_list.csv', usecols=(3, 4)):
+    corr_labels = loadtxt(mag_list, delimiter=',', usecols=usecols, dtype=str)
     channel = asarray(corr_labels[:, 1], dtype=int)-1
     corr_labels = [c.lower() for c in corr_labels[:, 0]]
+    return channel, corr_labels
+
+
+def import_response(filename='20_10_2017/fofb_rpm.txt'):
+    channel, corr_labels = generate_corr_labels()
 
     rsp_mat_raw = loadtxt(input_prefix+filename).T
 
@@ -28,7 +32,7 @@ def import_response(filename='20_10_2017/fofb_rpm.txt'):
     response_matrix[:,:,1] = rsp_mat_raw[channel, J:]
 
     response = Response(response_matrix, corr_labels, ['BPM%02i' % j for j in range(1, 55)],
-                        read_elemnames(input_prefix+'line.text'), unit='(mm/A)',
+                        read_elemnames(input_prefix+'line.text'), unit='mm/A',
                         corr_filters=('shk*', 'svk*'), drift_space=drift_info('u250'),
                         name=filename)
     response.pop_monitor('BPM12')
@@ -37,10 +41,7 @@ def import_response(filename='20_10_2017/fofb_rpm.txt'):
 
 if __name__=='__main__':
     save_path = 'delta_fofb_output/'
-    try:
-        makedirs(save_path)
-    except FileExistsError:
-        pass
+    makedir_if_nonexistent(save_path)
 
     response = import_response()